+++ example/lib/home_page.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:jocaagura_domain/jocaagura_domain.dart';
 
 import 'bloc_loading_demo_page.dart';
+import 'bloc_onboarding_demo_page.dart';
 import 'connectivity_demo_page.dart';
 import 'session_demo_page.dart';
 import 'ws_database_user_demo_page.dart';
@@ -45,6 +46,10 @@ class HomePage extends StatelessWidget {
           _NavigatorListTile(
             label: 'BlocLoading demo',
             page: BlocLoadingDemoPage(),
+          ),
+          _NavigatorListTile(
+            label: 'BlocOnboarding demo',
+            page: BlocOnboardingDemoPage(),
           ),
         ],
       ),